Moves GameState to 'model' package
