+++ app/src/main/java/com/example/bikesh/checkerz/model/IPlayer.java
@@ -1,7 +1,5 @@
 package com.example.bikesh.checkerz.model;
 
-import com.example.bikesh.checkerz.GameState;
-
 /**
  * Created by dhemard on 3/10/18.
  */
